% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assign_parameters}
\alias{assign_parameters}
\title{Assign the paramter values}
\usage{
assign_parameters(dt, req = NULL)
}
\arguments{
\item{dt}{data.table containing the following columns: parameter, value, and units.}

\item{req}{default set to NULL, if a string vector is read will check to make sure that
all of the required parameters are defined in the data table.}
}
\value{
Nothing, but has defined the elements of the parameter column as objects in
the environment.
}
\description{
Define parameters based on the a parameter data table that is read in.
}
