% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BAMS_config}
\alias{BAMS_config}
\title{The MEMC model configuration for BAMS}
\format{
An object created from \code{\link{memc_configure}}, containing the following elements.
\describe{
\item{name}{BAMS}
\item{table}{table of the dynamics used by this model configuration; see \link{dynamics} for more details.}
\item{params}{data table of the model parameters; see \link{memc_params} for more details.}
\item{state}{vector of the initial state values; see \link{memc_initial_state} for more details.}
}
}
\usage{
BAMS_config
}
\description{
The BAMS configuration is based off of Tang et al. 2022. This model
uses MM for the DOC uptake, MM for POC decomposition, and LM for MB
decay; see \link{dynamics} for more details.
}
\examples{
## Use memc_solve to run the simulation using this configuration.
print(BAMS_config)
memc_solve(mod = MEMS_config, time = 1:10)
}
\references{
Robertson et al.: Unifying soil organic matter formation and
persistence frameworks: the MEMS model, Biogeosciences, 16, 1225–1248, 2019.
\url{https://doi.org/10.5194/bg-16-1225-2019}
}
\seealso{
Other configurations: 
\code{\link{COMISSION_config}},
\code{\link{CORPSE_config}},
\code{\link{MEMS_config}},
\code{\link{MEND_config}},
\code{\link{MIMCS_config}}
}
\concept{configurations}
\keyword{datasets}
