% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_componets.R
\name{carbon_fluxes}
\alias{carbon_fluxes}
\title{Define the carbon fluxes.}
\usage{
carbon_fluxes(state, parms)
}
\arguments{
\item{state}{A numeric vector of the different carbon pool states.}

\item{parms}{A data frame of the parameters.}
}
\value{
A list of functions that calculate the fluxes between carbon pools.
}
\description{
\code{carbon_fluxes} Defines a system of equations that
describe fluxes between the general 5-pool scheme based on \href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}.
By retruning a list of named functions. The default configureation ise set up to follow the fluxes defined in \href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}.
More advanced users may choose to change the flux.
}
\references{
\href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}
}
\concept{carbon flux functions}
