% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{colorMEMCPalette}
\alias{colorMEMCPalette}
\title{Function that returns the MEMC color palette for the the default MEMC model configurations}
\usage{
colorMEMCPalette(name = NULL)
}
\arguments{
\item{name}{input vector containing the model names to return the color codes for, default will return colors for all the model configurations,}
}
\value{
vector containing color hex codes for the different model configurations
}
\description{
Function that returns the MEMC color palette for the the default MEMC model configurations
}
