% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COMISSION_model}
\alias{COMISSION_model}
\title{MEMC model configurations for MEND Wang et al. 2013}
\format{
an object created form \code{configure_model}.
\describe{
\item{name}{COMISSION}
\item{params}{the default MEMC parameter table}
\item{state}{the default inital state values for MEMC}
\item{carbon_pools_func}{carbon pools based on Wang et al. 2013}
\item{carbon_fluxes_func}{TBD}
}
}
\usage{
COMISSION_model
}
\description{
MEMC model configurations for MEND Wang et al. 2013
}
\examples{
COMISSION_model
solve_model(mod = COMISSION_model, time = 1:10)
}
\references{
\href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}

\href{https://doi.org/10.1890/12-0681.1}{update with the comission documetation}
}
\seealso{
Other MEMC configurations: 
\code{\link{MEND_model}}
}
\concept{MEMC configurations}
\keyword{datasets}
