% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_MEND}
\alias{compare_MEND}
\title{Solve different single P pool MEND configurations}
\usage{
compare_MEND(parameters, time, inital_state)
}
\arguments{
\item{parameters}{data.table containing the following columns: parameter, value, and units. Default values are stored as pacakge see \code{MEND_params}.}

\item{time}{a vector of the time setps.}

\item{inital_state}{a numeric vector of the different MEND carbon pool states. Default values are probvide as package data see \code{MEND_initalState}.}
}
\value{
a data frame of output from the different MEND configurations, \code{MEND}, \code{MEND_RM}, \code{MEND_ECA}, and \code{MEND_MILLEN}.
}
\description{
Solve different single P pool MEND configurations
}
\seealso{
Other intercomparison: 
\code{\link{compare_MEND2}()}

Other MEND: 
\code{\link{MEND_initalState}},
\code{\link{MEND_params}},
\code{\link{MEND}()}
}
\concept{MEND}
\concept{intercomparison}
