% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_MEND2}
\alias{compare_MEND2}
\title{Solve different two P pool MEND configurations}
\usage{
compare_MEND2(parameters, time, inital_state)
}
\arguments{
\item{parameters}{data.table containing the following columns: parameter, value, and units. Default values are stored as pacakge see \code{MEND2_params}.}

\item{time}{a vector of the time setps.}

\item{inital_state}{a numeric vector of the different MEND carbon pool states. Default values are probvide as package data see \code{MEND2_initalState}.}
}
\value{
a data frame of output from the different MEND configurations, \code{MEND2}, \code{MEND2_RM}, \code{MEND2_ECA}, and \code{MEND2_MILLEN}.
}
\description{
Solve different two P pool MEND configurations
}
\seealso{
Other intercomparison: 
\code{\link{compare_MEND}()}

Other MEND2: 
\code{\link{MEND2_initalState}},
\code{\link{MEND2_params}},
\code{\link{MEND2}()}
}
\concept{MEND2}
\concept{intercomparison}
