% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R
\name{configurations}
\alias{configurations}
\title{Configurations}
\description{
The various model configurations that are
shipped with the MEMC package, such as:
\code{MEND_model}, \code{COMISSION_model}, \code{CORPSE_model},
\code{MIMCS_model}, and \code{MEMS_model}.
}
\details{
These model configurations can be used with the \code{solve_model} function to
complete the desired model simulations.
}
\examples{
## Using solve_model with the MEND_model configuration.
solve_model(mode = MEND_model, time = 0:10)
}
\seealso{
dynamics
}
