% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{configure_model}
\alias{configure_model}
\title{Set up a model configuration}
\usage{
configure_model(
  params,
  state,
  carbon_pools_func,
  carbon_fluxes_func,
  name = NULL
)
}
\arguments{
\item{params}{data.table containing the following columns: parameter, value, and units.}

\item{state}{a vector of the intial state values, must be named}

\item{carbon_pools_func}{a function defining the carbon  pools}

\item{carbon_fluxes_func}{a function defining the carbon fluxes between pools}

\item{name}{default set to NULL otherwise is a string of the model configuration name.}
}
\description{
Set up a model configuration
}
\seealso{
Other helper functions: 
\code{\link{solve_model}()},
\code{\link{update_params}()}
}
\concept{helper functions}
