% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CORPSE_config}
\alias{CORPSE_config}
\title{The MEMC model configuration for CORPSE}
\format{
An object created from \code{\link{memc_configure}}, containing the following elements.
\describe{
\item{name}{CORPSE}
\item{table}{table of the dynamics used by this model configuration; see \link{dynamics} for more details.}
\item{params}{data table of the model parameters; see \link{memc_params} for more details.}
\item{state}{vector of the initial state values; see \link{memc_initial_state} for more details.}
}
}
\usage{
CORPSE_config
}
\description{
The CORPSE configuration is based off of Sulman et al. 2018. This
model uses RMM for DOC uptake, LM for POC decomposition, and LM for
MB decay; see \link{dynamics} for more details.
}
\examples{
## Use memc_solve to run the simulation using this configuration.
print(CORPSE_config)
memc_solve(mod = CORPSE_config, time = 1:10)
}
\references{
Sulman et al.: Multiple models and experiments underscore large
uncertainty in soil carbon dynamics, Biogeochemistry, 141, 109–123, 2018.
\url{https://doi.org/10.1007/s10533-018-0509-z}
}
\seealso{
Other configurations: 
\code{\link{BAMS_config}},
\code{\link{COMISSION_config}},
\code{\link{MEMS_config}},
\code{\link{MEND_config}},
\code{\link{MIMCS_config}}
}
\concept{configurations}
\keyword{datasets}
