% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R
\name{decay}
\alias{decay}
\title{Identifiers for MEMC supported decay}
\description{
These identifiers correspond to the types of decays users can select in \code{\link{carbon_fluxes}}.
}
\section{decay}{

\describe{
\item{DD}{Density dependent}
\item{LM}{Linear kinetics}
}
}

\references{
\href{doi}{ref}
}
\seealso{
carbon_fluxes
}
