% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R
\name{fluxes}
\alias{fluxes}
\title{Flexible SOM fluxes}
\description{
Description of the flexible SOM fluxes that users may change with \code{\link{configure_model}} function.
}
\details{
\describe{The MEMC package supports flexible definitions for the following SOM fluxes:
\item{DOMdecomp}{dissovled organic matter decomposition, the flux descirbing the DOM that is broken down for microbial biomoass uptake.}
\item{POMdecomp}{particulate oranic matter decomposition, the flux descirbing the POM that is broken down into DOM}
\item{MBdecay}{microbial biomass decay, the flux descirbing the micorbial biomass mortality/turn over}
}
}
\seealso{
\link{configurations}, fluxes, dynamics
}
