% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fme.R
\name{format_sensout}
\alias{format_sensout}
\title{Format the output returned by memc_sensrange or memc_sensfunc into long
data format for easy plotting}
\usage{
format_sensout(obj)
}
\arguments{
\item{obj}{object returned by memc_sensrange}
}
\value{
A long-format data frame containing the summary
from \code{FME}::\code{\link[FME]{sensRange}}.
}
\description{
Format the output returned by memc_sensrange or memc_sensfunc into long
data format for easy plotting
}
\examples{
\dontrun{
# Test the sensitivity of the MEND output for V.p, K.p, V.m, K.m, V.d, K.d
pars <- c("V.d" = 3.0e+00,"V.p" = 1.4e+01,"V.m" = 2.5e-01)
prange <- data.frame(min = pars - pars * 0.75,
max = pars + pars * 0.75)
t <- floor(seq(0, 365, length.out = 10))
out <- format_sensout(config = MEND_model, t = t, pars = pars,
parRange = prange, dist = "latin", num = 10)
to_plot <- format_sensout(out)
ggplot(data = to_plot) +
   geom_line(aes(time, Mean)) +
   geom_ribbon(aes(time, ymin = Min, ymax = Max), alpha = 0.5) +
   facet_wrap("variable", scales = "free")
}
}
\seealso{
Other sensitivity: 
\code{\link{memc_sensfunc}()},
\code{\link{memc_sensrange}()}
}
\concept{sensitivity}
