% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fme.R
\name{format_sensrange}
\alias{format_sensrange}
\title{Format the memc_sensrange output into long data format for easy plotting}
\usage{
format_sensrange(obj)
}
\arguments{
\item{obj}{object returned by memc_sensrange}
}
\value{
the a long dataframe of the summary memc_sensrange
}
\description{
Format the memc_sensrange output into long data format for easy plotting
}
\examples{
\dontrun{
# Test the sensitivity of the MEND output for V.p, K.p, V.m, K.m, V.d, K.d
pars <- c("V.d" = 3.0e+00,"V.p" = 1.4e+01,"V.m" = 2.5e-01)
prange <- data.frame(min = pars - pars * 0.75,
max = pars + pars * 0.75)
t <- floor(seq(0, 365, length.out = 10))
out <- memc_sensrange(config = MEND_model, t = t, pars = pars, parRange = prange, dist = "latin", num = 10)
to_plot <- format_sensrange(out)
}
}
