% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R
\name{kinetics}
\alias{kinetics}
\title{Identifiers for MEMC supported kinetics used}
\description{
These identifiers correspond to the various kinetics users can use in \code{\link{carbon_fluxes}} to
select for different kinetic dynamics.
}
\section{Kinetics}{

\describe{
\item{MM}{Michaelis–Menten kinetics, add ref}
\item{RMM}{Reverse Michaelis-Menten kinetics, add ref}
\item{ECA}{TODO}
\item{LM}{Linear kinetics model, add ref}
}
}

\concept{test}
