% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_fxns.R
\name{make_objective_function}
\alias{make_objective_function}
\title{Make the objective function}
\usage{
make_objective_function(mod, obs, x)
}
\arguments{
\item{mod}{memc model object created by \link{configure_model}}

\item{obs}{data.frame of the observations must contain values for MEMC state variables}

\item{x}{numeric vector with the values to be optimized, may be a model parameter or an initial state value}
}
\value{
function to be used as the objective function by mod cost
}
\description{
Internally used within \link{memc_modfit}
}
\seealso{
Other fit functions: 
\code{\link{memc_modcost}()},
\code{\link{memc_modfit}()}
}
\concept{fit functions}
