% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{memc_all_configs}
\alias{memc_all_configs}
\title{List of all the pre-defined MEMC models}
\format{
A easy-to-iterate-over list, each element of which is one of the MEMC models.
}
\usage{
memc_all_configs
}
\description{
List of all the pre-defined MEMC models
}
\examples{
memc_all_configs

# Return the summary table for dynamics used in model configurations
summary(memc_all_configs)
}
\seealso{
\code{\link{MEND_config}}, \code{\link{COMISSION_config}},
\code{\link{CORPSE_config}}, \code{\link{MEMS_config}},
\code{\link{BAMS_config}}, \code{\link{MIMCS_config}}
}
\keyword{datasets}
