% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-models.R
\docType{data}
\name{memc_all_models}
\alias{memc_all_models}
\title{List of all the pre-defined MEMC models}
\format{
A easy-to-iterate-over list, each element of which is one of the MEMC models.
}
\usage{
memc_all_models
}
\description{
List of all the pre-defined MEMC models
}
\examples{
print(memc_all_models)
}
\seealso{
\code{\link{MEND_model}}, \code{\link{COMISSION_model}},
\code{\link{CORPSE_model}}, \code{\link{MEMS_model}},
\code{\link{BAMS_model}}, \code{\link{MIMCS_model}}
}
\keyword{datasets}
