% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{memc_colorPalette}
\alias{memc_colorPalette}
\title{Return a color palette for the the default MEMC model configurations and other simulations}
\usage{
memc_colorPalette(name = NULL)
}
\arguments{
\item{name}{input vector containing the model names to return the color codes for, default will return colors for all the model configurations,}
}
\value{
A vector containing color hex codes for the different model configurations.
}
\description{
Return a color palette for the the default MEMC model configurations and other simulations
}
\examples{
\dontrun{
out <- memc_solve(MEND_config, time = 0:100)
ggplot(data = out) +
  geom_line(aes(time, value, color = name)) +
  scale_color_manual(values = memc_colorPalette()) + 
  facet_wrap("variable", scales = "free")
}
}
