% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{memc_configure}
\alias{memc_configure}
\title{Set up a model configuration}
\usage{
memc_configure(
  params,
  state,
  name = "unnamed",
  DOMuptake = "MM",
  POMdecomp = "MM",
  MBdecay = "LM"
)
}
\arguments{
\item{params}{data.table containing the following columns: parameter, value, and units.}

\item{state}{a vector of the initial state values, must be named}

\item{name}{string name of the model configuration, default set to "MEND".}

\item{DOMuptake}{string indicating the dynamics used to model microbial decomposition of DOM, one  of the following "MM", "RMM", or "ECA"}

\item{POMdecomp}{string indicating the dynamics used to model microbial decomposition of POM, one  of the following "MM", "RMM", "ECA", or "LM"}

\item{MBdecay}{string indicating microbial decay, one  of the following ""LM" or "DD"}
}
\value{
memc_single_model object of the name, dynamics, parameters and starting state values
}
\description{
Set up a model configuration
}
\examples{
# Modify the MEND model
m <- MEND_model
m_mod <- memc_configure(m$params, m$state, "MEND_modified", POMdecomp = "LM")
summary(m_mod)
memc_solve(m_mod, 0:10)
}
\seealso{
Other helper functions: 
\code{\link{memc_update_params}()},
\code{\link{update_state}()}
}
\concept{helper functions}
