% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{memc_data_andisol}
\alias{memc_data_andisol}
\alias{memc_data_gelisol}
\alias{memc_data_mollisol}
\alias{memc_data_oxisol}
\alias{memc_data_ultisol}
\title{Example soil incubation data from Wang et al. 2013}
\format{
A data frame of 45 rows and 3 columns containing respiration data
described in \href{doi.org/10.1890/12-0681.1}{Wang et al. 2013}.
\describe{
\item{Day}{Day of incubation.}
\item{Soil}{String constant describing soil type.}
\item{value}{Cumulative respiration.}
}

An object of class \code{data.frame} with 45 rows and 3 columns.

An object of class \code{data.frame} with 45 rows and 3 columns.

An object of class \code{data.frame} with 45 rows and 3 columns.

An object of class \code{data.frame} with 45 rows and 3 columns.
}
\usage{
data(memc_data_andisol)

memc_data_gelisol

memc_data_mollisol

memc_data_oxisol

memc_data_ultisol
}
\description{
Example soil incubation data from Wang et al. 2013
}
\examples{
memc_data_andisol
}
\references{
\href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}
}
\keyword{datasets}
