% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{memc_data_ultisol}
\alias{memc_data_ultisol}
\title{Soil incubation respiration fluxes from Wang et al. 2013}
\format{
A data frame of 45 rows (15 time points x three replicates)
and 3 columns containing respiration data.
\describe{
\item{Day}{Day of incubation.}
\item{Soil}{String constant describing soil type.}
\item{value}{Cumulative respiration, mg C/g soil.}
}
}
\usage{
data(memc_data_ultisol)
}
\description{
Soil incubation respiration fluxes from Wang et al. 2013
}
\examples{
memc_data_ultisol
}
\references{
\href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}
}
\keyword{datasets}
