% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{memc_incubation_gelisol}
\alias{memc_incubation_gelisol}
\title{Example data, time series and initial conditions for Gelisol from Wang et al. 2013 (10.1890/12-0681.1)}
\format{
A list of 2 elements:
\describe{
  \item{\code{data}}{data frame of  time-series data for carbon pools from the incubation experiment.
    The data frame contains the following variables:
    \describe{
      \item{\code{Day}}{numeric, indicating the time since the start of the incubation experiment.}
      \item{\code{Soil}}{Character indicating the soil type (e.g., "Gelisol").}
      \item{\code{IC}}{Numeric value for inorganic carbon (mg C g\eqn{^{-1}} soil).}
    }
  }
  \item{\code{state}}{named numeric vector of initial state values for the soil carbon pools.}
}
}
\usage{
memc_incubation_gelisol
}
\description{
A list containing incubation data and initial states.
}
\details{
The `data` element contains time-series observations, while the `state` element represents the initial conditions for the soil carbon pools.
}
\examples{
# Access the incubation data
memc_incubation_gelisol$data

# Access the initial state
memc_incubation_gelisol$state
}
\keyword{datasets}
