% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{memc_incubation_ultisol}
\alias{memc_incubation_ultisol}
\title{Soil incubation respiration fluxes from Wang et al. 2013}
\format{
A list containing:
\describe{
\item{data}{A data frame with 45 rows (15 time points x three replicates) and 3 columns:
\describe{
\item{Day}{Day of incubation.}
\item{Soil}{String constant describing soil type.}
\item{value}{Cumulative respiration, mg C/g soil.}
}}
\item{state}{A named vector of the initial states of the carbon pools associated with the observations}
}

A list of 2 elements:
\describe{
  \item{\code{data}}{data frame of  time-series data for carbon pools from the incubation experiment.
    The data frame contains the following variables:
    \describe{
      \item{\code{Day}}{numeric, indicating the time since the start of the incubation experiment.}
      \item{\code{Soil}}{Character indicating the soil type (e.g., "Ultisol").}
      \item{\code{IC}}{Numeric value for inorganic carbon (mg C g\eqn{^{-1}} soil).}
    }
  }
  \item{\code{state}}{named numeric vector of initial state values for the soil carbon pools.}
}
}
\usage{
data(memc_incubation_ultisol)

memc_incubation_ultisol
}
\description{
A list containing incubation data and initial states.
}
\details{
The `data` element contains time-series observations, while the `state` element represents the initial conditions for the soil carbon pools.
}
\examples{
memc_incubation_ultisol
# Access the incubation data
memc_incubation_ultisol$data

# Access the initial state
memc_incubation_ultisol$state
}
\references{
\href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}
}
\keyword{datasets}
