% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_fxns.R
\name{memc_modcost}
\alias{memc_modcost}
\title{Calculate the model cost between output and obs data frame}
\usage{
memc_modcost(output, obs)
}
\arguments{
\item{output}{data.frame returned by}

\item{obs}{data.frame of observations or the data to fit the model to}
}
\value{
modCost object
}
\description{
Internally used within \link{make_objective_function}, this is a MEMC wrapper for \link[pkg:FME]{modCost}.
}
\seealso{
Other fit functions: 
\code{\link{make_objective_function}()},
\code{\link{memc_modfit}()}
}
\concept{fit functions}
