% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{memc_models}
\alias{memc_models}
\title{Get a list of all available models in MEMC}
\usage{
memc_models()
}
\value{
A list of models that is easy to iterate over.
}
\description{
Get a list of all available models in MEMC
}
\examples{
memc_models()
}
\seealso{
\code{\link{MEND_model}}, \code{\link{COMISSION_model}},
\code{\link{CORPSE_model}}, \code{\link{MEMS_model}},
\code{\link{BAMS_model}}, \code{\link{MIMCS_model}}
}
