% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_fxns.R
\name{memc_modfit}
\alias{memc_modfit}
\title{Constrain MEMC model to observational data}
\usage{
memc_modfit(mod, obs, x, ...)
}
\arguments{
\item{mod}{memc model object created by \link{configure_model}}

\item{obs}{data.frame of the observations must contain values for MEMC state variables}

\item{x}{numeric vector with the values to be optimized, may be a model parameter or an initial state value, must be named}

\item{...}{additional arguments passed to \link[pkg:FME]{modCost}}
}
\value{
function to be used as the objective function by mod cost
}
\description{
Constrain MEMC model to observational data
}
\seealso{
Other fit functions: 
\code{\link{make_objective_function}()},
\code{\link{memc_modcost}()}
}
\concept{fit functions}
