% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{memc_params}
\alias{memc_params}
\title{The parameter values for the basic MEMC model configurations from Wang et al. 2013}
\format{
A data frame of 20 rows and 4 columns containing the model
parameter values from \href{doi.org/10.1890/12-0681.1}{Wang et al. 2013}.
\describe{
\item{parameter}{String indicating MEMC parameter.}
\item{description}{String describing the parameter.}
\item{units}{String of the parameter units.}
\item{value}{Parameter values from Table 2 of Wang et al. 2013}
}
}
\usage{
data(memc_params)
}
\description{
The parameter values for the basic MEMC model configurations from Wang et al. 2013
}
\examples{
memc_params
}
\references{
\href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}
}
\seealso{
Other inputs: 
\code{\link{memc_initial_state}}
}
\concept{inputs}
\keyword{datasets}
