% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{memc_plot}
\alias{memc_plot}
\title{Helper function that quickly visualizes results returned by memc_solve, memc_sensfunc, or memc_sensrange}
\usage{
memc_plot(obj)
}
\arguments{
\item{obj}{object to plot, data frame returned by memc_solve or the object created by memc_sensfunc, or memc_sensrange}
}
\value{
a ggplot2 plot
}
\description{
Helper function that quickly visualizes results returned by memc_solve, memc_sensfunc, or memc_sensrange
}
