% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fme.R
\name{memc_sensfun}
\alias{memc_sensfun}
\title{Local sensitivity for a MEMC model}
\usage{
memc_sensfun(config, t, x, ...)
}
\arguments{
\item{config}{a MEMC model configuration object, either one of the pre-built configurations listed in \code{memc_all_configs} or created using \code{memc_configure}}

\item{t}{vector of the time steps to run the model at}

\item{x}{vector of the parameters or initial state values to test}

\item{...}{additional arguments passed to \code{\link[FME]{sensFun}}}
}
\value{
The results of \code{FME}::\code{\link[FME]{sensFun}}.
}
\description{
Estimate the local effect of a parameter on a MEMC model output
}
\examples{
\dontrun{
# Test the sensitivity of the MEND output for V_p, K_p, V_m
pars <- c("V_d" = 3.0e+00,"V_p" = 1.4e+01,"V_m" = 2.5e-01)
out <- memc_sensfun(config = MEND_config, t = t, x = pars)
plot(out)
}
}
\seealso{
Other sensitivity: 
\code{\link{memc_sensrange}()}
}
\concept{sensitivity}
