% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fme.R
\name{memc_sensfunc}
\alias{memc_sensfunc}
\title{Estimate the local effect of a parameter on a memc model output}
\usage{
memc_sensfunc(config, t, pars, ...)
}
\arguments{
\item{config}{memc model configuration}

\item{t}{vector of the time steps to run the model at}

\item{pars}{vector of the parameters to test in FME::sensFun}

\item{...}{additional arguments passed to FME::sensFun}
}
\value{
the results of the FME::sensFun
}
\description{
Estimate the local effect of a parameter on a memc model output
}
\examples{
\dontrun{
# Test the sensitivity of the MEND output for V.p, K.p, V.m
pars <- c("V.d" = 3.0e+00,"V.p" = 1.4e+01,"V.m" = 2.5e-01)
out <- memc_sensfunc(config = MEND_model, t = t, pars = pars)
pairs(out)
plot(out)
# Using the helper functions to make nic ggplots
to_plot <- format_sensout(out)
ggplot(data = to_plot) +
   geom_line(aes(time, value, color = parameter)) +
   facet_wrap("variable", scales = "free")
}
}
\seealso{
Other sensitivity: 
\code{\link{format_sensout}()},
\code{\link{memc_sensrange}()}
}
\concept{sensitivity}
