% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fme.R
\name{memc_sensrange}
\alias{memc_sensrange}
\title{Run the memc model with the FME sensRange}
\usage{
memc_sensrange(config, t, pars, parRange, dist, ...)
}
\arguments{
\item{config}{memc model configuration}

\item{t}{vector of the time steps to run the model at}

\item{pars}{vector of the parameters to test in sensRange}

\item{parRange}{data frame of the min/max parameter values}

\item{dist}{str for the distribution according to which the parameters will be sampled}

\item{...}{additional arguments passed to FME::sensRange}
}
\value{
the results of the FME::sensRange
}
\description{
Run the memc model with the FME sensRange
}
\examples{
\dontrun{
# Test the sensitivity of the MEND output for V.p, K.p, V.m
pars <- c("V.d" = 3.0e+00,"V.p" = 1.4e+01,"V.m" = 2.5e-01)
prange <- data.frame(min = pars - pars * 0.75,
max = pars + pars * 0.75)
t <- floor(seq(0, 365, length.out = 10))
out <- memc_sensrange(config = MEND_model, t = t, pars = pars,
parRange = prange, dist = "latin", num = 10)
plot(summary(out))
# Using the helper functions.
to_plot <- format_sensout(out)
ggplot(data = to_plot) +
   geom_line(aes(time, Mean)) +
   geom_ribbon(aes(time, ymin = Min, ymax = Max), alpha = 0.5) +
   facet_wrap("variable", scales = "free")
}
}
\seealso{
Other sensitivity: 
\code{\link{format_sensout}()},
\code{\link{memc_sensfunc}()}
}
\concept{sensitivity}
