% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_component.R
\name{memc_solve}
\alias{memc_solve}
\title{Solve a MEMC configuration}
\usage{
memc_solve(mod, time, params = NULL, state = NULL, ...)
}
\arguments{
\item{mod}{model object, for example one of the list entries returned
by \code{\link{memc_all_models}}}

\item{time}{a numeric vector of the time steps to solve over}

\item{params}{default set to NULL, will then use the parameter table
read in with the \code{mod} object}

\item{state}{default set to NULL, will then use the state read read
in with the \code{mod} object}

\item{...}{additional arguments passed to \code{\link[deSolve]{ode}}}
}
\value{
A long-formatted \code{\link[data.table]{data.table}} of the
simulation results; \code{time} = hour.
}
\description{
Solve a MEMC configuration
}
\concept{helper}
