% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{memc_update_config}
\alias{memc_update_config}
\title{Update a model configuration}
\usage{
memc_update_config(mod, new = NULL)
}
\arguments{
\item{mod}{a MEMC model configuration object created by \code{memc_configure}}

\item{new}{vector containing the parameters and or initial pool values}
}
\description{
Update a model configuration
}
\seealso{
Other helper functions: 
\code{\link{memc_configure}()},
\code{\link{memc_update_params}()},
\code{\link{update_state}()}
}
\concept{helper functions}
