% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{memc_update_params}
\alias{memc_update_params}
\title{Update the parameter table with new values}
\usage{
memc_update_params(new_params, param_table)
}
\arguments{
\item{new_params}{a named vector of parameter values to update the param table.}

\item{param_table}{data.table containing the following columns: parameter, value, and units, this is the basic parameter table that will be updated with the new values.}
}
\value{
Updated \code{\link[data.table]{data.table}} containing the new parameter values.
}
\description{
Update the parameter table with new values
}
\seealso{
Other helper functions: 
\code{\link{memc_configure}()},
\code{\link{memc_update_config}()},
\code{\link{update_state}()}
}
\concept{helper functions}
\concept{parameters}
