% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MEMS_config}
\alias{MEMS_config}
\title{The MEMC model configuration for MEMS}
\format{
An object created from \code{\link{memc_configure}}, containing the following elements.
\describe{
\item{name}{MEMS}
\item{table}{table of the dynamics used by this model configuration; see \link{dynamics} for more details.}
\item{params}{data table of the model parameters; see \link{memc_params} for more details.}
\item{state}{vector of the initial state values; see \link{memc_initial_state} for more details.}
}
}
\usage{
MEMS_config
}
\description{
The MEMS configuration is based off of Robertson et al. 2019. This
model uses LM for the DOM uptake, LM for POM decomposition, and LM
for MB decay; see \link{dynamics} for more details.
}
\examples{
## Use memc_solve to run the simulation using this configuration.
print(MEMS_config)
memc_solve(mod = MEMS_config, time = 1:10)
}
\references{
\href{https://doi.org/10.5194/bg-16-1225-2019}{Robertson et al. 2019}
}
\seealso{
Other configurations: 
\code{\link{BAMS_config}},
\code{\link{COMISSION_config}},
\code{\link{CORPSE_config}},
\code{\link{MEND_config}},
\code{\link{MIMCS_config}}
}
\concept{configurations}
\keyword{datasets}
