% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEND2013.R
\name{MEND2013}
\alias{MEND2013}
\title{Solve MEND 2013}
\usage{
MEND2013(parameters, time, inital_state)
}
\arguments{
\item{parameters}{data.table containing the following columns: parameter, value, and units}

\item{time}{a vector of the time setps}

\item{intial_state}{A numeric vector of the different MEND carbon pool states.}
}
\value{
a data frame of MEND output variables
}
\description{
\code{MEND2013} Run and solve MEND 2013  \href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}.
}
\seealso{
Other 2013 MEND model functions: 
\code{\link{MEND2013_fluxes}()},
\code{\link{MEND2013_pools}()}
}
\concept{2013 MEND model functions}
\concept{model}
