% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MEND2013_params}
\alias{MEND2013_params}
\title{The parameter values for MEND 2013}
\format{
A data table consisting of 4 columns with inputs from \href{doi.org/10.1890/12-0681.1}{Wang et al. 2013}.
\describe{
\item{parameter}{String character of the default MEND parameters.}
\item{description}{String character describing the parameter.}
\item{units}{String character of the parameter units.}
\item{value}{Numeric values taken from the table 2 of Wang et al. 2013}
}
}
\usage{
MEND2013_params
}
\description{
The parameter values for MEND 2013
}
\references{
\href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}
}
\seealso{
Other parameters: 
\code{\link{MEND2013_initalState}}
}
\concept{MEND2012}
\concept{parameters}
\keyword{datasets}
