% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MEND2_initalState}
\alias{MEND2_initalState}
\title{The inital carbon pool values for MEND 2013}
\format{
A vector of the intial states of the Carbon pools from MEND 2013 \href{doi.org/10.1890/12-0681.1}{Wang et al. 2013}.
\describe{ a named vecotr of the MEND 2013 Carbon pools
\item{'P1'}{'First pool of particulate organic carbon or POC'}
\item{'P2'}{'Second pool of particulate organic carbon or POC'}
\item{'M'}{'Mineral associated organic carbon or MOC, note that this excludes Q'}
\item{'Q1'}{'Adsorbed phase from DOC pool 1'}
\item{'Q2'}{'Adsorbed phase from DOC pool 2'}
\item{'B'}{'Microbial biomass carbon or MBC'}
\item{'D1'}{'Dissolved organic carbon or DOC from pool 1'}
\item{'D2'}{'Dissolved organic carbon or DOC from pool 2'}
\item{'EP1'}{'Enzyme for decomposition of P1'}
\item{'EP2'}{'Enzyme for decomposition of P2'}
\item{'EM'}{'Ensyme for decomposition of M'}
\item{'IC'}{'Accumulated CO2 flux'}
\item{'Tot'}{'Total carbon in the patch'}
}
}
\usage{
MEND2_initalState
}
\description{
The inital carbon pool values for MEND 2013
}
\references{
\href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}
}
\seealso{
Other inital values: 
\code{\link{MEND_initalState}}

Other MEND2: 
\code{\link{MEND2_params}},
\code{\link{MEND2}()},
\code{\link{compare_MEND2}()}
}
\concept{MEND2}
\concept{inital values}
\keyword{datasets}
