% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEND2_MILLEN.R
\name{MEND2_MILLEN}
\alias{MEND2_MILLEN}
\title{Solve 2 P pool MEND with ECA kinetics}
\usage{
MEND2_MILLEN(parameters, time, inital_state, method = NULL)
}
\arguments{
\item{parameters}{data.table containing the following columns: parameter, value, and units. Default values are stored as pacakge see \code{MEND2_params}.}

\item{time}{a vector of the time setps.}

\item{inital_state}{a numeric vector of the different MEND carbon pool states. Default values are probvide as package data see \code{MEND2_initalState}.}

\item{method}{optional string argument that can be passed in as the method into the \code{ode}.}
}
\value{
a data frame of MEND output variables
}
\description{
\code{MEND2_ECA} Run and solve 2 P pool MEND  \href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013} that
uses ECA kinetics in microbial uptake of DOC.
}
\examples{
# define the time vector & load the default parameters and initial state values stored
# as package data.
t <- seq(0, 175200, 24) ##per hour 24hour by 365 days =8760  20year=175200
p <- MEND2_params
state <- MEND2_initalState

# Solve MEND2_MILLEN
out <- MEND2_MILLEN(parameters = p, time = t,  inital_state = state)
}
\seealso{
Other models: 
\code{\link{MEND2_ECA}()},
\code{\link{MEND2_RM}()},
\code{\link{MEND2}()},
\code{\link{MEND_ECA}()},
\code{\link{MEND}()}
}
\concept{MEND2_MIlLEN}
\concept{models}
