% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MEND2_params}
\alias{MEND2_params}
\title{The parameter values for the 2 pool MEND}
\format{
A data table consisting of 4 columns with inputs from \href{TODO}{TODO}.
\describe{
\item{parameter}{String character of the default 2 pool MEND parameters.}
\item{description}{String character describing the parameter.}
\item{units}{String character of the parameter units.}
\item{value}{Numeric values taken from TODO}
}
}
\usage{
MEND2_params
}
\description{
The parameter values for the 2 pool MEND
}
\references{
\href{TODO}{TODO}
}
\seealso{
Other parameters: 
\code{\link{MEND_params}}

Other MEND2: 
\code{\link{MEND2_initalState}},
\code{\link{MEND2}()},
\code{\link{compare_MEND2}()}
}
\concept{MEND2}
\concept{parameters}
\keyword{datasets}
