% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MEND_config}
\alias{MEND_config}
\title{The MEMC model configuration for MEND}
\format{
An object created from \code{\link{memc_configure}}.
\describe{
\item{name}{MEND}
\item{table}{table of the dynamics used by this model configuration; see \link{dynamics} for more details.}
\item{params}{data table of the model parameters; see \link{memc_params} for more details.}
\item{state}{vector of the initial state values; see \link{memc_initial_state} for more details.}
}
}
\usage{
MEND_config
}
\description{
The MEND configuration is based off of Wang et al. 2013. This
configuration uses MM for DOM uptake, MM for POM decomposition,
and LM for MB decay; see \link{dynamics} for more details.
}
\examples{
## Use memc_solve to run the simulation using this configuration.
print(MEND_config)
memc_solve(mod = MEND_config, time = 1:10)
}
\references{
Wang et al.: Development of microbial-enzyme-mediated
decomposition model parameters through steady-state and dynamic analyses,
Ecol. Appl., 23, 255–272, 2013. \url{https://doi.org/10.1890/12-0681.1}
}
\seealso{
Other configurations: 
\code{\link{BAMS_config}},
\code{\link{COMISSION_config}},
\code{\link{CORPSE_config}},
\code{\link{MEMS_config}},
\code{\link{MIMCS_config}}
}
\concept{configurations}
\keyword{datasets}
