% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEND_ECA.R
\name{MEND_ECA}
\alias{MEND_ECA}
\title{Solve MEND with ECA kinetics}
\usage{
MEND_ECA(parameters, time, inital_state)
}
\arguments{
\item{parameters}{data.table containing the following columns: parameter, value, and units. Default values are stored as pacakge see \code{MEND_params}.}

\item{time}{a vector of the time setps.}

\item{inital_state}{a numeric vector of the different MEND carbon pool states. Default values are probvide as package data see \code{MEND_initalState}}
}
\value{
a data frame of MEND output variables
}
\description{
\code{MEND_ECA} Run and solve MEND 2013  \href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013} that
uses ECA kinetics in microbial uptake of DOC.
}
\examples{
# define the time vector & load the default parameters and initial state values stored
# as package data.
t <- seq(0, 175200, 24) ##per hour 24hour by 365 days =8760  20year=175200
p <- MEND_params
state <- MEND_initalState

# Solve MEND_ECA
out <- MEND_ECA(parameters = p, time = t,  inital_state = state)
}
\seealso{
Other models: 
\code{\link{MEND2_ECA}()},
\code{\link{MEND2_MILLEN}()},
\code{\link{MEND2_RM}()},
\code{\link{MEND2}()},
\code{\link{MEND}()}
}
\concept{MEND_ECA}
\concept{models}
