% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MEND_initalState}
\alias{MEND_initalState}
\title{The inital carbon pool values for MEND 2013}
\format{
A vector of the intial states of the Carbon pools from MEND 2013 \href{doi.org/10.1890/12-0681.1}{Wang et al. 2013}.
\describe{ a named vecotr of the MEND 2013 Carbon pools
\item{'P'}{'Particulate organic carbon or POC'}
\item{'M'}{'Mineral associated organic carbon or MOC, note that this excludes Q'}
\item{'Q'}{'Adsorbed phase of DOC'}
\item{'B'}{'Microbial biomass carbon or MBC'}
\item{'D'}{'Dissolved organic carbon or DOC'}
\item{'EP'}{'Enzyme for decomposition of P'}
\item{'EM'}{'Ensyme for decomposition of M'}
\item{'IC'}{'Accumulated CO2 flux'}
\item{'Tot'}{'Total carbon in the patch'}
}
}
\usage{
MEND_initalState
}
\description{
The inital carbon pool values for MEND 2013
}
\references{
\href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}
}
\seealso{
Other inital values: 
\code{\link{MEND2_initalState}}

Other MEND: 
\code{\link{MEND_params}},
\code{\link{MEND}()},
\code{\link{compare_MEND}()}
}
\concept{MEND}
\concept{inital values}
\keyword{datasets}
