% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEND_RM.R
\name{MEND_RM}
\alias{MEND_RM}
\title{Solve MEND 2013 with reverse michaelis menten kinetics}
\usage{
MEND_RM(parameters, time, inital_state)
}
\arguments{
\item{parameters}{data.table containing the following columns: parameter, value, and units. Default values are stored as pacakge see \code{MEND_params}.}

\item{time}{a vector of the time setps.}

\item{inital_state}{a numeric vector of the different MEND carbon pool states. Default values are probvide as package data see \code{MEND_initalState}}
}
\value{
a data frame of MEND output variables
}
\description{
\code{MEND_RM} Run and solve MEND 2013  \href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013} that
uses reverse michaelis menten kinetics in microbial uptake of DOC.
}
\concept{MEND_RM}
\concept{models
# define the time vector & load the default parameters and initial state values stored
# as package data.
t <- seq(0, 175200, 24) ##per hour 24hour by 365 days =8760  20year=175200
p <- MEND_params
state <- MEND_initalState

# Solve MEND_RM
out <- MEND_RM(parameters = p, time = t,  inital_state = state)}
