% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MIMCS_config}
\alias{MIMCS_config}
\title{The MEMC model configuration for MIMCS}
\format{
An object created from \code{\link{memc_configure}}, containing the following elements.
\describe{
\item{name}{MIMCS}
\item{table}{table of the dynamics used by this model configuration; see \link{dynamics} for more details.}
\item{params}{data table of the model parameters; see \link{memc_params} for more details.}
\item{state}{vector of the initial state values; see \link{memc_initial_state} for more details.}
}
}
\usage{
MIMCS_config
}
\description{
The MIMCS configuration is based off of Wieder et al. 2015. This
model uses MM for the DOC uptake, MM for POC decomposition, and DD
for MB decay; see \link{dynamics} for more details.
}
\examples{
## Use memc_solve to run the simulation using this configuration.
print(MIMCS_config)
memc_solve(mod = MIMCS_config, time = 1:10)
}
\references{
Wieder, W. R., et al.: Explicitly representing soil microbial
processes in Earth system models, Global Biogeochem. Cycles, 29, 1782–1800,
2015. \url{https://doi.org/10.1002/2015GB005188}
}
\seealso{
Other configurations: 
\code{\link{BAMS_config}},
\code{\link{COMISSION_config}},
\code{\link{CORPSE_config}},
\code{\link{MEMS_config}},
\code{\link{MEND_config}}
}
\concept{configurations}
\keyword{datasets}
