% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MIMCS_model}
\alias{MIMCS_model}
\title{The MEMC model configuration for MIMCS}
\format{
an object created form \code{configure_model}, containing the following elements.
\describe{
\item{name}{MIMCS}
\item{table}{table of the dynamics used by this model configuration, see \link{dynamics} for more details.}
\item{params}{data table of the model parameters, see \link{default_params} for more details.}
\item{state}{vector of the initial state values, see \link{default_initial} for more details.}
}
}
\usage{
MIMCS_model
}
\description{
The MIMCS configuration is based off of Wieder et al. 2015, this model uses MM for the DOM
uptake, MM for POM decomposition, and a DD for MB decay, see \link{dynamics} for more
details.
}
\examples{
## Use solve_model to run the simulation using this configuration.
print(MIMCS_model)
solve_model(mod = MIMCS_model, time = 1:10)
}
\references{
\href{https://doi.org/10.1002/2015GB005188}{Wieder et al. 2015}
}
\seealso{
Other {configurations}: 
\code{\link{BAMS_model}},
\code{\link{COMISSION_model}},
\code{\link{CORPSE_model}},
\code{\link{MEMS_model}},
\code{\link{MEND_model}}
}
\concept{{configurations}}
\keyword{datasets}
