% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{modify_fluxes_func}
\alias{modify_fluxes_func}
\title{Modify a flux functions}
\usage{
modify_fluxes_func(flux_func, F1 = "MM")
}
\arguments{
\item{flux_func}{a function that will return a list of functions that modify how carbon moves between the carbon pools.}

\item{F1}{default set to MM, could be one of "MM", "RMM", "ECA", "LM", each referring to a type of kinetics}
}
\value{
a list of carbon flux functions
}
\description{
Modify a flux functions
}
\seealso{
Other helper functions: 
\code{\link{configure_model}()},
\code{\link{solve_model}()},
\code{\link{update_params}()}
}
\concept{helper functions}
