% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_class_fxns.R
\name{plot.memc_sensFunc}
\alias{plot.memc_sensFunc}
\title{Quick plot the results returned from memc_sensFunc}
\usage{
\method{plot}{memc_sensFunc}(x, y, ...)
}
\arguments{
\item{x}{An object of class `memc_sensFunc`.}

\item{y}{Ignored}

\item{...}{additional arguments}
}
\description{
Quick plot the results returned from memc_sensFunc
}
