% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_class_fxns.R
\name{plot.memc_sensRange}
\alias{plot.memc_sensRange}
\title{Quick plot the results returned from memc_sensRange}
\usage{
\method{plot}{memc_sensRange}(x, y, ...)
}
\arguments{
\item{x}{An object of class `memc_sensRange`.}

\item{y}{Ignore}

\item{...}{additional arguments center, lower, upper that may be used to define the line and ribbon boundaries.}
}
\description{
Quick plot the results returned from memc_sensRange
}
