% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_class_fxns.R
\name{plot.memc_solve}
\alias{plot.memc_solve}
\title{Quick plot the results returned from memc_solve}
\usage{
\method{plot}{memc_solve}(x, y, ...)
}
\arguments{
\item{x}{An object of class `memc_solve`.}

\item{y}{Ignore}

\item{...}{Additional arguments (ignored).}
}
\description{
Quick plot the results returned from memc_solve
}
