% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{print.memc_single_config}
\alias{print.memc_single_config}
\title{Return a full memc model object}
\usage{
\method{print}{memc_single_config}(x, ...)
}
\arguments{
\item{x}{An object of class `memc_single_config`.}

\item{...}{Additional arguments (ignored).}
}
\description{
Provides details of all single model configuration
}
