% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{print.single_model}
\alias{print.single_model}
\title{Return a full memc model object}
\usage{
\method{print}{single_model}(x, ...)
}
\arguments{
\item{x}{An object of class `single_model`.}

\item{...}{Additional arguments (ignored).}
}
\description{
Provides details of all single model configuration
}
