% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_class_fxns.R
\name{rbind.memc_solve}
\alias{rbind.memc_solve}
\title{Combine results returned from memc_solve}
\usage{
\method{rbind}{memc_solve}(..., deparse.level = 1)
}
\arguments{
\item{...}{the `memc_solve` objects to be combined into a single data frame}

\item{deparse.level}{the default deparse.level = 1 typically and deparse.level = 2 always construct labels from the argument names, see the ‘Value’ section below.}
}
\description{
Combine results returned from memc_solve
}
