% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_functions.R
\name{sm_format_out}
\alias{sm_format_out}
\title{Solve a MEMC configuration}
\usage{
sm_format_out(rslt, mod)
}
\arguments{
\item{rslt}{deSolve matrix}

\item{mod}{model object created by \code{make_model}}
}
\value{
a long formatted data.table of the simulation results
}
\description{
Solve a MEMC configuration
}
\seealso{
Other helper functions: 
\code{\link{configure_model}()},
\code{\link{sm_internal}()},
\code{\link{update_params}()}
}
\concept{helper functions}
