% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{solver}
\alias{solver}
\title{MEMC solver function}
\usage{
solver(params, time, state, carbon_pools_func, carbon_fluxes_func, ...)
}
\arguments{
\item{params}{data.table containing the following columns: parameter, value, and units.}

\item{time}{a vector of the time setps}

\item{state}{a vector of the intial state values, must be named}

\item{carbon_pools_func}{a function defining the carbon  pools}

\item{carbon_fluxes_func}{a function defining the carbon fluxes between pools}

\item{...}{additional arguments that can be read into \code{deSolve::ode}}
}
\value{
a long formatted data.table of the simulation results
}
\description{
Based on user defined carbon pools and flux functions solve a system of equations.
}
\concept{helper function}
