% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{summary.memc_all_configs}
\alias{summary.memc_all_configs}
\title{Summary table of memc_all_configs}
\usage{
\method{summary}{memc_all_configs}(object, ...)
}
\arguments{
\item{object}{An object of class `memc_all_configs`.}

\item{...}{Additional arguments (ignored).}
}
\description{
Provides a summary of dynamics used in `memc_all_configs`.
}
