% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{summary.memc_single_config}
\alias{summary.memc_single_config}
\title{Summary table for a single memc model}
\usage{
\method{summary}{memc_single_config}(object, ...)
}
\arguments{
\item{object}{An object of class `memc_single_config`.}

\item{...}{Additional arguments (ignored).}
}
\description{
Provides a summary of dynamics used in `memc_single_config`.
}
