% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{update_state}
\alias{update_state}
\title{TODO update}
\usage{
update_state(new_vals, state)
}
\arguments{
\item{new_vals}{update}

\item{state}{update}
}
\description{
TODO update
}
\seealso{
Other helper functions: 
\code{\link{update_params}()}
}
\concept{helper functions}
