% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_componets.R
\name{carbon_fluxes}
\alias{carbon_fluxes}
\title{Define the carbon fluxes.}
\usage{
carbon_fluxes(F1 = "MM", env = NULL, state = NULL, params = NULL)
}
\arguments{
\item{F1}{string indicating the type of enzyme kinetics used in the microbial decomposition of POM.}

\item{env}{an environment of the initial state and parameter values created by \code{internal_load_params}, by default it is set to an empty environment.}

\item{state}{A numeric vector of the different carbon pool states that will be used to up date the preset values read in from the env, default set to NULL will use the entries in the env object.}

\item{params}{A data frame of the parameters that will be used to up date the entries in the env environment, by default set to NULL.}
}
\value{
A list of functions that calculate the fluxes between carbon pools.
}
\description{
\code{carbon_fluxes} Defines a system of equations that
describe fluxes between the general 5-pool scheme based on \href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}.
By returning a list of named functions. The default configuration is set up to follow the fluxes defined in \href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}.
More advanced users may choose to change the flux.
}
\references{
\href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}
}
\concept{carbon flux functions}
