% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{colorMEMCPalette}
\alias{colorMEMCPalette}
\title{Return the MEMC color palette for the the default MEMC model configurations}
\usage{
colorMEMCPalette(name = NULL)
}
\arguments{
\item{name}{input vector containing the model names to return the color codes for, default will return colors for all the model configurations,}
}
\value{
A vector containing color hex codes for the different model configurations.
}
\description{
Return the MEMC color palette for the the default MEMC model configurations
}
\examples{
\dontrun{
out <- memc_solve(model = MEND_model, time = 0:100)
ggplot(data = out, aes(time, value)) +
geom_line(color = name) +
scale_color_manual(values = colorMEMCPalette("MEND")) +
facet_wrap("variable")
}
}
