% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COMISSION_model}
\alias{COMISSION_model}
\title{MEMC model configurations for MEND Wang et al. 2013}
\format{
an object created form \code{configure_model}.
\describe{
\item{name}{COMISSION, the model name}
\item{table}{table describing the user selected kinetics}
\item{params}{the default MEMC parameter table}
\item{state}{the default inital state values for MEMC}
}
}
\usage{
COMISSION_model
}
\description{
MEMC model configurations for MEND Wang et al. 2013
}
\examples{
COMISSION_model
solve_model(mod = COMISSION_model, time = 1:10)
}
\seealso{
Other models: 
\code{\link{CORPSE_model}},
\code{\link{MEND_model}}
}
\concept{models}
\keyword{datasets}
