% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COMISSION_model}
\alias{COMISSION_model}
\title{MEMC COMISSION model configuration based on  \href{https://doi.org/10.1016/j.soilbio.2020.107912}{Ahrens et al. 2020}}
\format{
an object created form \code{configure_model}.
\describe{
\item{name}{COMISSION}
\item{params}{the default MEMC parameter table}
\item{state}{the default inital state values for MEMC}
\item{carbon_pools_func}{carbon pools based on Wang et al. 2013}
\item{carbon_fluxes_func}{carbon fluxes based on Wang et al. 2013. The COMISSION_model MEMC
configuration uses RMM kinetics for DOMdecomp, MM for  POMdecomp, and LM for MBdecay.
see \code{\link{kinetics}} for more details.}
}
}
\usage{
COMISSION_model
}
\description{
MEMC COMISSION model configuration based on  \href{https://doi.org/10.1016/j.soilbio.2020.107912}{Ahrens et al. 2020}
}
\examples{
COMISSION_model
solve_model(mod = COMISSION_model, time = 1:10)
}
\references{
\href{https://doi.org/10.1016/j.soilbio.2020.107912}{Ahrens et al. 2020}

\href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}
}
\seealso{
Other MEMC configurations: 
\code{\link{CORPSE_model}},
\code{\link{MEND_model}}
}
\concept{MEMC configurations}
\keyword{datasets}
