% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R
\name{configurations}
\alias{configurations}
\title{Configurations}
\description{
The various model configurations that are
shipped with the MEMC package, such as:
\code{BAMS_config}, \code{COMISSION_config},
\code{CORPSE_config}, \code{MEND_config},
\code{MIMCS_config}, and \code{MEMS_config}.
All supported model configurations are listed
in \code{memc_all_configs}.
}
\details{
These model configurations can be used with the \code{memc_solve} function to
complete the desired model simulations.
}
\examples{
## Using memc_solve with the MEND_config configuration.
out <- memc_solve(mod = MEND_config, time = 0:10)
}
\seealso{
dynamics, fluxes
}
