% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R
\name{configurations}
\alias{configurations}
\title{Configurations}
\description{
The various model configurations that are
shipped with the MEMC package, such as:
\code{BAMS_model}, \code{COMISSION_model},
\code{CORPSE_model}, \code{MEND_model},
\code{MIMCS_model}, and \code{MEMS_model}.
All supported model configurations are listed
in \code{memc_all_models}.
}
\details{
These model configurations can be used with the \code{memc_solve} function to
complete the desired model simulations.
}
\examples{
## Using memc_solve with the MEND_model configuration.
out <- memc_solve(mod = MEND_model, time = 0:10)
}
\seealso{
dynamics, fluxes
}
