% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{configure_model}
\alias{configure_model}
\title{Set up a model configuration}
\usage{
configure_model(
  params,
  state,
  carbon_pools_func = carbon_pools,
  carbon_fluxes_func = carbon_fluxes,
  name = "MEND",
  DOMdecomp = "MM",
  POMdecomp = "MM",
  MBdecay = "LM"
)
}
\arguments{
\item{params}{data.table containing the following columns: parameter, value, and units.}

\item{state}{a vector of the initial state values, must be named}

\item{carbon_pools_func}{a function defining the carbon pools, by default is set to the \code{carbon_pools}}

\item{carbon_fluxes_func}{a function defining the carbon fluxes between pools, by default is set to \code{carbon_fluxes}}

\item{name}{string name of the model configuration, default set to "MEND".}

\item{DOMdecomp}{string indicating the type of enzyme kinetics used in the microbial decomposition of DOM, one  of the following "MM", "RMM", "ECA", or "LM", see \code{\link{kinetics}} for more details.}

\item{POMdecomp}{string indicating the type of enzyme kinetics used in the microbial decomposition of POM, one  of the following "MM", "RMM", "ECA", or "LM", see \code{\link{kinetics}} for more details.}

\item{MBdecay}{string indicating microbial decay, one  of the following "MM", "RMM", "ECA", or "LM", see \code{\link{kinetics}} for more details.}
}
\description{
Set up a model configuration
}
\seealso{
Other helper functions: 
\code{\link{sm_format_out}()},
\code{\link{sm_internal}()},
\code{\link{update_params}()}
}
\concept{helper functions}
