% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{configure_model}
\alias{configure_model}
\title{Set up a model configuration}
\usage{
configure_model(
  params,
  state,
  carbon_pools_func = carbon_pools,
  carbon_fluxes_func,
  name = "MEND",
  F1 = "MM"
)
}
\arguments{
\item{params}{data.table containing the following columns: parameter, value, and units.}

\item{state}{a vector of the initial state values, must be named}

\item{carbon_pools_func}{a function defining the carbon pools, by default is set to the \code{carbon_pools}}

\item{carbon_fluxes_func}{a function defining the carbon fluxes between pools, by default is set to \code{carbon_fluxes}}

\item{name}{default set to MEND, is the string name of the model.}

\item{F1}{string for the F1 carbon flux microbial kinetics, by default set to MM ("Michaelis Menten")}

\item{config}{knitr_kable describing the model configuration}
}
\description{
Set up a model configuration
}
\seealso{
Other helper functions: 
\code{\link{modify_fluxes_func}()},
\code{\link{solve_model}()},
\code{\link{update_params}()}
}
\concept{helper functions}
