% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CORPSE_model}
\alias{CORPSE_model}
\title{MEMC CORPSE model configuration based on \href{https://doi.org/10.1038/nclimate2436}{Sulman et al. 2014}}
\format{
an object created by \code{configure_model}.
\describe{
\item{name}{CORPSE}
\item{params}{the default MEMC parameter table}
\item{state}{the default inital state values for MEMC}
\item{carbon_pools_func}{carbon pools based on Wang et al. 2013}
\item{carbon_fluxes_func}{carbon fluxes based on Wang et al. 2013. The CORPSE_model MEMC
configuration uses RMM kinetics for DOMdecomp, LM for POMdecomp, and LM for MBdecay.
see \code{\link{kinetics}} for more details.}
}
}
\usage{
CORPSE_model
}
\description{
MEMC CORPSE model configuration based on \href{https://doi.org/10.1038/nclimate2436}{Sulman et al. 2014}
}
\examples{
CORPSE_model
solve_model(mod = CORPSE_model, time = 1:10)
}
\references{
\href{doi:10.1038/nclimate2436}{Sulman, B. N., R. P. Phillips, A. C. Oishi, E. Shevliakova, and S. W. Pacala (2014), Microbe-driven turnover offsets mineral-mediated storage of soil carbon under elevated CO2, Nat. Clim. Change, 4(12), 1099–1102, doi:10.1038/nclimate2436.}

\href{https://doi.org/10.1890/12-0681.1}{Wang, G., Post, W.M. and Mayes, M.A. (2013), Development of microbial-enzyme-mediated decomposition model parameters through steady-state and dynamic analyses. Ecological Applications, 23: 255-272. https://doi.org/10.1890/12-0681.1}
}
\seealso{
Other MEMC configurations: 
\code{\link{COMISSION_model}},
\code{\link{MEND_model}}
}
\concept{MEMC configurations}
\keyword{datasets}
