% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CORPSE_model}
\alias{CORPSE_model}
\title{MEMC CORPSE model configuration based on \href{https://doi.org/10.1038/nclimate2436}{Sulman et al. 2014}}
\format{
an object created form \code{configure_model}.
\describe{
\item{name}{CORPSE}
\item{params}{the default MEMC parameter table}
\item{state}{the default inital state values for MEMC}
\item{carbon_pools_func}{carbon pools based on Wang et al. 2013}
\item{carbon_fluxes_func}{carbon fluxes based on Wang et al. 2013. The CORPSE_model MEMC
configuration uses RMM kinetics for DOMdecomp, LM for POMdecomp, and LM for MBdecay.
see \code{\link{kinetics}} for more details.}
}
}
\usage{
CORPSE_model
}
\description{
MEMC CORPSE model configuration based on \href{https://doi.org/10.1038/nclimate2436}{Sulman et al. 2014}
}
\examples{
CORPSE_model
solve_model(mod = CORPSE_model, time = 1:10)
}
\references{
\href{https://doi.org/10.1038/nclimate2436}{Sulman et al. 2014}

\href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}
}
\seealso{
Other MEMC configurations: 
\code{\link{COMISSION_model}},
\code{\link{MEND_model}}
}
\concept{MEMC configurations}
\keyword{datasets}
