% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CORPSE_model}
\alias{CORPSE_model}
\title{The MEMC model configuration for CORPSE}
\format{
An object created from \code{\link{memc_configure}}, containing the following elements.
\describe{
\item{name}{CORPSE}
\item{table}{table of the dynamics used by this model configuration; see \link{dynamics} for more details.}
\item{params}{data table of the model parameters; see \link{memc_params} for more details.}
\item{state}{vector of the initial state values; see \link{default_initial} for more details.}
}
}
\usage{
CORPSE_model
}
\description{
The CORPSE configuration is based off of Sulman et al. 2018. This
model uses RMM for DOM uptake, LM for POM decomposition, and LM for
MB decay; see \link{dynamics} for more details.
}
\examples{
## Use memc_solve to run the simulation using this configuration.
print(CORPSE_model)
memc_solve(mod = CORPSE_model, time = 1:10)
}
\references{
\href{https://doi.org/10.1007/s10533-018-0509-z}{Sulman et al. 2018}
}
\seealso{
Other configurations: 
\code{\link{BAMS_model}},
\code{\link{COMISSION_model}},
\code{\link{MEMS_model}},
\code{\link{MEND_model}},
\code{\link{MIMCS_model}},
\code{\link{model_configs}}
}
\concept{configurations}
\keyword{datasets}
