% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CORPSE_model}
\alias{CORPSE_model}
\title{MEMC model configurations}
\format{
an object created form \code{configure_model}.
\describe{
\item{name}{CORPSE, the model name}
\item{table}{table describing the user selected kinetics}
\item{params}{the default MEMC parameter table}
\item{state}{the default inital state values for MEMC}
}
}
\usage{
CORPSE_model
}
\description{
MEMC model configurations
}
\examples{
CORPSE_model
solve_model(mod = CORPSE_model, time = 1:10)
}
\seealso{
Other models: 
\code{\link{COMISSION_model}},
\code{\link{MEND_model}}
}
\concept{models}
\keyword{datasets}
