% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_initial}
\alias{default_initial}
\title{The default initial carbon pool values used by the default MEMC model configurations}
\format{
A named vector of the initial states of the carbon pools.
\describe{ a named vector of the Carbon pools
\item{POM}{Particulate organic carbon or POC}
\item{MOM}{Mineral associated organic carbon or MOC, note that this excludes QOM}
\item{QOM}{Adsorbed phase of DOC}
\item{MB}{Microbial biomass carbon or MBC}
\item{DOM}{Dissolved organic carbon or DOC}
\item{EP}{Enzyme for decomposition of POM}
\item{EM}{Enzyme for decomposition of MOM}
\item{IC}{Accumulated CO2 flux}
\item{Tot}{Total carbon in the patch}
}
}
\usage{
default_initial
}
\description{
The default initial carbon pool values used by the default MEMC model configurations
}
\examples{
print(default_initial)
}
\references{
\href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}
}
\seealso{
Other inputs: 
\code{\link{memc_params}}
}
\concept{inputs}
\keyword{datasets}
