% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_params}
\alias{default_params}
\title{The parameter values for the basic MEMC model configurations from Wang et al. 2013}
\format{
A data table consisting of 4 columns with inputs from \href{doi.org/10.1890/12-0681.1}{Wang et al. 2013}.
}
\usage{
data(default_params)
}
\description{
These identifiers correspond to the various kinetics users can use in \code{\link{carbon_fluxes}} to
select for different kinetic dynamics.
}
\section{Columns}{

\describe{
\item{parameter}{String character of the default MEND parameters.}
\item{description}{String character describing the parameter.}
\item{units}{String character of the parameter units.}
\item{value}{Numeric values taken from the table 2 of Wang et al. 2013}
}
}

\references{
\href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}
}
\keyword{datasets}
