% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_params}
\alias{default_params}
\title{Default parameter values used in MEMC model configurations}
\format{
A data table consisting of 4 columns with inputs
}
\usage{
data(default_params)
}
\description{
Default parameter values used in MEMC model configurations,
from \href{doi.org/10.1890/12-0681.1}{Wang et al. 2013}.
}
\section{Columns}{

\describe{
\item{parameter}{String character of the default MEND parameters.}
\item{description}{String character describing the parameter.}
\item{units}{String character of the parameter units.}
\item{value}{Numeric values taken from the table 2 of Wang et al. 2013}
}
}

\references{
\href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}
}
\seealso{
Other parameters: 
\code{\link{update_params}()}
}
\concept{parameters}
\keyword{datasets}
