% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_params}
\alias{default_params}
\title{The parameter values for the basic MEMC model configurations from Wang et al. 2013}
\format{
A data frame of 20 rows and 4 columns containing the model parameter values from \href{doi.org/10.1890/12-0681.1}{Wang et al. 2013}.
\describe{
\item{parameter}{String character of the default MEND parameters.}
\item{description}{String character describing the parameter.}
\item{units}{String character of the parameter units.}
\item{value}{Numeric values taken from the table 2 of Wang et al. 2013}
}
}
\usage{
data(default_params)
}
\description{
The parameter values for the basic MEMC model configurations from Wang et al. 2013
}
\examples{
print(default_params)
}
\references{
\href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}
}
\seealso{
Other inputs: 
\code{\link{default_initial}}
}
\concept{inputs}
\keyword{datasets}
