% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R
\name{fluxes}
\alias{fluxes}
\title{Flexible SOM fluxes}
\description{
Description of the flexible SOM fluxes that users may change with \code{\link{memc_configure}} function.
}
\details{
\describe{The MEMC package supports flexible definitions for the following SOM fluxes:
\item{F1}{dissolved organic uptake by microbial biomass, the flux describing the DOM that is broken down for microbial biomoass uptake.}
\item{POMdecomp}{particulate oranic matter decomposition, the flux describing the POM that is broken down into DOM.}
\item{MBdecay}{microbial biomass decay, the flux descirbing the micorbial biomass mortality/turn over.}
}
}
\seealso{
configurations, fluxes, dynamics
}
