% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R
\name{kinetics}
\alias{kinetics}
\title{Identifiers for MEMC supported kinetics used}
\description{
These identifiers correspond to the various kinetics users can use in \code{\link{carbon_fluxes}} to
select for different kinetic dynamics.
}
\section{Kinetics}{

\describe{
\item{MM}{Michaelis–Menten kinetics, TODO add ref}
\item{RMM}{Reverse Michaelis-Menten kinetics, TODO add ref}
\item{ECA}{Equilibrium chemistry approximation, based on RESOM (Tang and Riley, 2014)}
\item{LM}{Linear kinetics model, TODO add ref}
}
}

\references{
\href{https://doi.org/10.1038/nclimate2438}{Tang and Riley, 2014}
}
\concept{test}
