% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_data.R
\name{make_objective_function}
\alias{make_objective_function}
\title{Make the objective function to use in \link[FME]{modFit}}
\usage{
make_objective_function(obs, config)
}
\arguments{
\item{obs}{data.frame of the observations to fit the model to}

\item{config}{MEMC configuration}
}
\value{
function to be used a the f argument in
}
\description{
Make the objective function to use in \link[FME]{modFit}
}
