% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{memc_incubation_ultisol}
\alias{memc_incubation_ultisol}
\title{Soil incubation respiration fluxes from Wang et al. 2013}
\format{
A list containing:
\describe{
\item{data}{A data frame with 45 rows (15 time points x three replicates) and 3 columns:
\describe{
\item{Day}{Day of incubation.}
\item{Soil}{String constant describing soil type.}
\item{value}{Cumulative respiration, mg C/g soil.}
}}
\item{state}{A named vector of the initial states of the carbon pools associated with the observations}
}
}
\usage{
data(memc_incubation_ultisol)
}
\description{
Soil incubation respiration fluxes from Wang et al. 2013
}
\examples{
memc_incubation_ultisol
}
\references{
\href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}
}
\keyword{datasets}
