% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fme.R
\name{memc_modfit}
\alias{memc_modfit}
\title{Fit a MEMC model to a comparison data}
\usage{
memc_modfit(config, x, comp_data, lower = -Inf, upper = Inf, ...)
}
\arguments{
\item{config}{memc model configuration}

\item{x}{vector of the parameters or state values to be fit using FME::modFit}

\item{comp_data}{data frame containing the comparison to fit the data to}

\item{lower}{lower bounds on the parameters; if unbounded set equal to -Inf}

\item{upper}{bounds on the parameters; if unbounded set equal to Inf}

\item{...}{addition arguments that may be passed to FME::modFit}
}
\value{
the results of the FME::modFit
}
\description{
Fit a MEMC model to a comparison data
}
\seealso{
Other fme: 
\code{\link{memc_sensrange}()}
}
\concept{fme}
