% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{memc_params}
\alias{memc_params}
\title{Parameter values for the MEMC model configuration}
\format{
A data frame of 20 rows and 4 columns containing the default MEMC
parameter values
\describe{
\item{parameter}{String indicating MEMC parameter.}
\item{description}{String describing the parameter.}
\item{units}{String of the parameter units.}
\item{value}{Parameter values}
}
}
\usage{
data(memc_params)
}
\description{
Parameter values for the MEMC model configuration
}
\examples{
memc_params
}
\references{
Wang et al.: Development of microbial-enzyme-mediated
decomposition model parameters through steady-state and dynamic analyses,
Ecol. Appl., 23, 255–272, 2013. \url{https://doi.org/10.1890/12-0681.1}
}
\seealso{
Other inputs: 
\code{\link{memc_initial_state}}
}
\concept{inputs}
\keyword{datasets}
