% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{memc_params}
\alias{memc_params}
\title{The parameter values for the basic MEMC model configurations from Wang et al. 2013}
\format{
A data frame of 20 rows and 4 columns containing the default MEMC
parameter values
\describe{
\item{parameter}{String indicating MEMC parameter.}
\item{description}{String describing the parameter.}
\item{units}{String of the parameter units.}
\item{value}{Parameter values}
}
}
\usage{
data(memc_params)
}
\description{
The parameter values for the basic MEMC model configurations from Wang et al. 2013
}
\examples{
memc_params
}
\seealso{
Other inputs: 
\code{\link{memc_initial_state}}
}
\concept{inputs}
\keyword{datasets}
