% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fme.R
\name{memc_sensrange}
\alias{memc_sensrange}
\title{Global sensitivity ranges for a MEMC model}
\usage{
memc_sensrange(config, t, x, parRange, dist, ...)
}
\arguments{
\item{config}{a MEMC model configuration object, either one of the pre-built configurations listed in \code{memc_all_configs} or created using \code{memc_configure}}

\item{t}{vector of the time steps to run the model at}

\item{x}{vector of the parameters or initial model pool sizes that will be varied}

\item{parRange}{data frame of the min/max parameter values}

\item{dist}{str for the distribution according to which the parameters will be sampled from, options" "unif" (uniformly random samples), "norm", (normally distributed random samples), "latin" (latin hypercube distribution), and "grid" (parameters arranged on a grid).}

\item{...}{additional arguments passed to \code{\link[FME]{sensRange}}}
}
\value{
The results of \code{FME}::\code{\link[FME]{sensRange}}.
}
\description{
Given a MEM model configuration, estimate the global parameter sensitivity.
}
\examples{
\dontrun{
# Test the sensitivity of the MEND output for V_p, K_p, V_m
pars <- c("V_d" = 3.0e+00,"V_p" = 1.4e+01,"V_m" = 2.5e-01)
prange <- data.frame(min = pars - pars * 0.75,
max = pars + pars * 0.75)
t <- floor(seq(0, 365, length.out = 10))
out <- memc_sensrange(config = MEND_config, t = t, x = pars,
parRange = prange, dist = "latin", num = 10)
plot(out)
}
}
\seealso{
Other sensitivity: 
\code{\link{memc_sensfun}()}
}
\concept{sensitivity}
