% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_component.R
\name{memc_solve}
\alias{memc_solve}
\title{Solve a MEMC configuration}
\usage{
memc_solve(mod, time, params = NULL, state = NULL, ...)
}
\arguments{
\item{mod}{model object, typically one of the \code{\link{memc_all_models}}
or modified from \code{\link{memc_configure}}}

\item{time}{a vector of the time steps}

\item{params}{default set to NULL, will then use the parameter table
read in with the \code{mod} object}

\item{state}{default set to NULL, will then use the state read read
in with the \code{mod} object}

\item{...}{additional arguments passed to \code{\link[deSolve]{ode}}}
}
\value{
A long-formatted \code{\link[data.table]{data.table}} of the
simulation results, with columns:
\item{time}{Time point, hours}
\item{variable}{Name of model carbon pool}
\item{value}{Value of the pool}
\item{units}{Units of the pool value}
\item{name}{Model name used}
}
\description{
Solve a MEMC configuration
}
\examples{
memc_solve(MEND_model, time = 0:10)
}
\seealso{
\code{\link{memc_configure}}
}
\concept{helper}
