% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MEND_model}
\alias{MEND_model}
\title{MEMC model configurations for MEND Wang et al. 2013}
\format{
an object created form \code{configure_model}.
\describe{
\item{name}{MEND, the model name}
\item{table}{table describing the user selected kinetics}
\item{params}{the default MEMC parameter table}
\item{state}{the default inital state values for MEMC}
}
}
\usage{
MEND_model
}
\description{
MEMC model configurations for MEND Wang et al. 2013
}
\examples{
MEND_model
solve_model(mod = MEND_model, time = 1:10)
}
\seealso{
Other MEMC configurations: 
\code{\link{COMISSION_model}},
\code{\link{CORPSE_model}}
}
\concept{MEMC configurations}
\keyword{datasets}
