% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MEND_model}
\alias{MEND_model}
\title{The MEMC model configuration for MEND}
\format{
an object created form \code{configure_model}.
\describe{
\item{name}{MEND}
\item{table}{table of the dynamics used by this model configuration, see \link{dynamics} for more details.}
\item{params}{data table of the model parameters, see \link{default_params} for more details.}
\item{state}{vector of the initial state values, see \link{default_initial} for more details.}
}
}
\usage{
MEND_model
}
\description{
The MEND configuration is based off of Wang et al. 2013, this
configuration uses MM for DOM uptake, MM for
POM decomposition, and a LM for MB decay, see \link{dynamics} for more
details.
}
\examples{
## Use solve_model to run the simulation using this configuration.
print(MEND_model)
solve_model(mod = MEND_model, time = 1:10)
}
\references{
\href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}
}
\seealso{
Other {configurations}: 
\code{\link{BAMS_model}},
\code{\link{COMISSION_model}},
\code{\link{CORPSE_model}},
\code{\link{MEMS_model}},
\code{\link{MIMCS_model}},
\code{\link{model_configs}}
}
\concept{{configurations}}
\keyword{datasets}
