% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MEND_model}
\alias{MEND_model}
\title{The MEMC model configuration for MEND}
\format{
An object created from \code{\link{memc_configure}}.
\describe{
\item{name}{MEND}
\item{table}{table of the dynamics used by this model configuration; see \link{dynamics} for more details.}
\item{params}{data table of the model parameters; see \link{memc_params} for more details.}
\item{state}{vector of the initial state values; see \link{memc_initial_state} for more details.}
}
}
\usage{
MEND_model
}
\description{
The MEND configuration is based off of Wang et al. 2013. This
configuration uses MM for DOM uptake, MM for POM decomposition,
and LM for MB decay; see \link{dynamics} for more details.
}
\examples{
## Use memc_solve to run the simulation using this configuration.
print(MEND_model)
memc_solve(mod = MEND_model, time = 1:10)
}
\references{
\href{https://doi.org/10.1890/12-0681.1}{Wang et al. 2013}
}
\seealso{
Other configurations: 
\code{\link{BAMS_model}},
\code{\link{COMISSION_model}},
\code{\link{CORPSE_model}},
\code{\link{MEMS_model}},
\code{\link{MIMCS_model}}
}
\concept{configurations}
\keyword{datasets}
