% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-models.R
\docType{data}
\name{MIMCS_model}
\alias{MIMCS_model}
\title{The MEMC model configuration for MIMCS}
\format{
An object created from \code{\link{memc_configure}}, containing the following elements.
\describe{
\item{name}{MIMCS}
\item{table}{table of the dynamics used by this model configuration; see \link{dynamics} for more details.}
\item{params}{data table of the model parameters; see \link{default_params} for more details.}
\item{state}{vector of the initial state values; see \link{memc_state_ultisol} for more details.}
}
}
\usage{
MIMCS_model
}
\description{
The MIMCS configuration is based off of Wieder et al. 2015. This
model uses MM for the DOM uptake, MM for POM decomposition, and DD
for MB decay; see \link{dynamics} for more details.
}
\examples{
## Use memc_solve to run the simulation using this configuration.
print(MIMCS_model)
memc_solve(mod = MIMCS_model, time = 1:10)
}
\references{
\href{https://doi.org/10.1002/2015GB005188}{Wieder et al. 2015}
}
\seealso{
Other configurations: 
\code{\link{BAMS_model}},
\code{\link{COMISSION_model}},
\code{\link{CORPSE_model}},
\code{\link{MEMS_model}},
\code{\link{MEND_model}},
\code{\link{model_configs}}
}
\concept{configurations}
\keyword{datasets}
