% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_configs}
\alias{model_configs}
\title{Data table of all the pre-defined MEMC model configurations}
\format{
data table containing 4 columns
\describe{
\item{model}{model configuration name}
\item{DOMuptake}{DOM uptake by microbial biomass; see \link{dynamics} for more details.}
\item{POMdecomp}{POM decomposition; see \link{dynamics} for more details.}
\item{MBdecay}{microbial biomass decay; see \link{default_initial} for more details.}
}
}
\usage{
model_configs
}
\description{
Data table of all the pre-defined MEMC model configurations
}
\examples{
print(model_configs)
}
\seealso{
Other configurations: 
\code{\link{BAMS_model}},
\code{\link{COMISSION_model}},
\code{\link{CORPSE_model}},
\code{\link{MEMS_model}},
\code{\link{MEND_model}},
\code{\link{MIMCS_model}}
}
\concept{configurations}
\keyword{datasets}
