% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{update_state}
\alias{update_state}
\title{Update the state vector values this is for internal function use}
\usage{
update_state(new_vals, state)
}
\arguments{
\item{new_vals}{update}

\item{state}{update}
}
\description{
Update the state vector values this is for internal function use
}
\seealso{
Other helper functions: 
\code{\link{configure_model}()},
\code{\link{update_params}()}
}
\concept{helper functions}
